+++ src/components/ES6/simple-3D-background-three2/js/shader/fragment-custom2.glsl
@@ -1,0 +1,17 @@
+uniform float time;
+
+varying vec2 vUv;
+
+void main( void ) {
+
+    vec2 position = vUv;
+
+    float color = 0.0;
+    color += sin( position.x * cos( time / 15.0 ) * 80.0 ) + cos( position.y * cos( time / 15.0 ) * 10.0 );
+    color += sin( position.y * sin( time / 10.0 ) * 40.0 ) + cos( position.x * sin( time / 25.0 ) * 40.0 );
+    color += sin( position.x * sin( time / 5.0 ) * 10.0 ) + sin( position.y * sin( time / 35.0 ) * 80.0 );
+    color *= sin( time / 10.0 ) * 0.5;
+
+    gl_FragColor = vec4( vec3( color, color * 0.5, sin( color + time / 3.0 ) * 0.75 ), 1.0 );
+
+}